+++ jetty-http/src/main/java/org/eclipse/jetty/http/pathmap/PathSpec.java
@@ -18,6 +18,14 @@
 
 package org.eclipse.jetty.http.pathmap;
 
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.Objects;
+import java.util.stream.Stream;
+
+import org.eclipse.jetty.util.ArrayUtil;
+import org.eclipse.jetty.util.annotation.ManagedAttribute;
+
 /**
  * A path specification is a URI path template that can be matched against.
  * <p>
@@ -86,9 +94,9 @@ public interface PathSpec extends Comparable<PathSpec>
     String getSuffix();
 
     /**
-     *  Test if a normalized string pathspec is equivalent to this pathspec
-     * @param pathSpec String representation to be normalized
-     * @return true of the string represents the same pathspec
+     *  Test if a string is equivalent to this pathspec
+     * @param pathSpec String representation to be normalized and then compared to the {@link #getDeclaration()} string.
+     * @return true if the string, after normalization, represents the same pathspec
      */
     boolean is(String pathSpec);
 
@@ -99,4 +107,151 @@ public interface PathSpec extends Comparable<PathSpec>
      * @return true if the path matches this path spec, false otherwise
      */
     boolean matches(String path);
+
+    /**
+     * A Mapping of one or more PathSpecs
+     */
+    class Mapping
+    {
+        private PathSpec[] _pathSpecs = {}; // never null
+
+        /**
+         * @param pathSpec The pathSpec to set, which is assumed to be a {@link ServletPathSpec}
+         */
+        public void addPathSpec(String pathSpec)
+        {
+            Objects.requireNonNull(pathSpec);
+            addPathSpec(new ServletPathSpec(pathSpec));
+        }
+
+        /**
+         * @param pathSpec The pathSpec to set, which may be an arbitrary type of pathspec
+         */
+        public void addPathSpec(PathSpec pathSpec)
+        {
+            Objects.requireNonNull(pathSpec);
+            _pathSpecs = ArrayUtil.addToArray(_pathSpecs, pathSpec, PathSpec.class);
+        }
+
+        /**
+         * Test if the list of path specs contains a particular one.
+         *
+         * @param pathSpec the path spec
+         * @return true if path spec matches something in mappings
+         */
+        public boolean containsPathSpec(String pathSpec)
+        {
+            return stream().anyMatch(p -> p.is(pathSpec));
+        }
+
+        /**
+         * Test if the list of path specs contains a particular one.
+         *
+         * @param pathSpec the path spec
+         * @return true if path spec matches something in mappings
+         */
+        public boolean containsPathSpec(PathSpec pathSpec)
+        {
+            return stream().anyMatch(p -> p.equals(pathSpec));
+        }
+
+        /**
+         * @return Returns only the {@link ServletPathSpec}s as strings or empty array.
+         * @deprecated Use {@link #getServletPathSpecs()}
+         */
+        @Deprecated
+        public String[] getPathSpecs()
+        {
+            return getServletPathSpecs();
+        }
+
+        /**
+         * @return Returns only the {@link ServletPathSpec}s as strings or empty array.
+         */
+        @ManagedAttribute(value = "servlet url patterns", readonly = true)
+        public String[] getServletPathSpecs()
+        {
+            return Arrays.stream(_pathSpecs)
+                .filter(ServletPathSpec.class::isInstance)
+                .map(PathSpec::getDeclaration)
+                .toArray(String[]::new);
+        }
+
+        public boolean hasPathSpecs()
+        {
+            return _pathSpecs.length > 0;
+        }
+
+        /**
+         * @param pathSpec The pathSpec to set, which is assumed to be a {@link ServletPathSpec}
+         */
+        public void setPathSpec(String pathSpec)
+        {
+            Objects.requireNonNull(pathSpec);
+            _pathSpecs = new PathSpec[]{new ServletPathSpec(pathSpec)};
+        }
+
+        /**
+         * @param pathSpec The pathSpec to set
+         */
+        public void setPathSpec(PathSpec pathSpec)
+        {
+            Objects.requireNonNull(pathSpec);
+            _pathSpecs = new PathSpec[]{pathSpec};
+        }
+
+        /**
+         * @param pathSpecs The pathSpecs to set, which are assumed to be {@link ServletPathSpec}s
+         * @deprecated Use {@link #setServletPathSpecs(String[])}
+         */
+        @Deprecated
+        public void setPathSpecs(String[] pathSpecs)
+        {
+            setServletPathSpecs(pathSpecs);
+        }
+
+        /**
+         * @param pathSpecs The pathSpecs to set, which are assumed to be {@link ServletPathSpec}s
+         */
+        public void setPathSpecs(PathSpec[] pathSpecs)
+        {
+            _pathSpecs = (pathSpecs == null)
+                ? new PathSpec[]{}
+                : Arrays.stream(pathSpecs).filter(Objects::nonNull).toArray(PathSpec[]::new);
+        }
+
+        /**
+         * @param pathSpecs The pathSpecs to set, which are assumed to be {@link ServletPathSpec}s
+         */
+        public void setPathSpecs(Collection<PathSpec> pathSpecs)
+        {
+            _pathSpecs = (pathSpecs == null)
+                ? new PathSpec[]{}
+                : pathSpecs.stream().filter(Objects::nonNull).toArray(PathSpec[]::new);
+        }
+
+        /**
+         * @param pathSpecs The pathSpecs to set, which are assumed to be {@link ServletPathSpec}s
+         */
+        public void setServletPathSpecs(String[] pathSpecs)
+        {
+            _pathSpecs = (pathSpecs == null)
+                ? new PathSpec[]{}
+                : Arrays.stream(pathSpecs).filter(Objects::nonNull).map(ServletPathSpec::new).toArray(PathSpec[]::new);
+        }
+
+        public Stream<PathSpec> stream()
+        {
+            return Arrays.stream(_pathSpecs);
+        }
+
+        /**
+         * @return Returns the pathSpecs as array of {@link PathSpec} instances or empty array.
+         */
+        @ManagedAttribute(value = "all path spec patterns", readonly = true)
+        public PathSpec[] toPathSpecs()
+        {
+            return Arrays.copyOf(_pathSpecs, _pathSpecs.length);
+        }
+    }
 }
